third_party/WebKit/LayoutTests/inspector/console/console-log-eval-syntax-error-expected.txt

CONSOLE ERROR: line 1: Uncaught SyntaxError: Unexpected token }
CONSOLE ERROR: line 2: Uncaught SyntaxError: Unexpected token }
Tests that syntax errors in eval are logged into console, contains correct link and doesn't cause browser crash.

VM:1 Uncaught SyntaxError: Unexpected token }
    at foo (console-log-eval-syntax-error.html:7)
foo @ console-log-eval-syntax-error.html:7
setTimeout (async)
(anonymous) @ VM:1
boo.js:2 Uncaught SyntaxError: Unexpected token }
    at boo (console-log-eval-syntax-error.html:10)
boo @ console-log-eval-syntax-error.html:10
setTimeout (async)
(anonymous) @ VM:1
